move and enable profiler_tests again
