+++ crates/iota-replay/Cargo.toml
@@ -51,9 +51,11 @@ move-core-types.workspace = true
 move-vm-config.workspace = true
 
 [dev-dependencies]
+iota-execution = { workspace = true, features = ["tracing"] }
 iota-macros.workspace = true
 iota-metrics.workspace = true
 iota-simulator.workspace = true
+move-vm-profiler.workspace = true
 test-cluster.workspace = true
 
 [[example]]
